code 10: Documentation of the Mobility catalog.
