+++ src/main/java/com/project/app_dog_back/application/mapper/IMovilidadMapper.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.application.mapper;
 
-import com.project.app_dog_back.application.dto.MovilidadDto;
+import com.project.app_dog_back.application.dto.catalog.MovilidadDto;
 import com.project.app_dog_back.domain.model.catalog.Movilidad;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
