+++ src/main/java/com/project/app_dog_back/domain/service/IMovilidadService.java
@@ -1,6 +1,6 @@
 package com.project.app_dog_back.domain.service;
 
-import com.project.app_dog_back.application.dto.MovilidadDto;
+import com.project.app_dog_back.application.dto.catalog.MovilidadDto;
 import com.project.app_dog_back.domain.model.catalog.Movilidad;
 
 import java.util.List;
